requirements.txt
beautifulsoup4==4.13.4
Flask==3.1.1
langchain==0.3.26
langchain_community==0.3.27
langchain_core==0.3.68
langchain_openai==0.3.27
pandas==2.3.1
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.4
twilio==9.6.3